requirements.txt
libtiff==0.4.2
numpy==1.16.4
Pillow==6.0.0
PyQt5==5.12.1
PyQt5-sip==4.19.17
PyQt5-stubs==5.12.1.0
